requirement.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
aiohttp @ file:///C:/ci/aiohttp_1632749616620/work
async-timeout==3.0.1
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1625831763871/work
chardet @ file:///C:/ci/chardet_1607706912142/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.0.3
cryptography @ file:///C:/ci/cryptography_1616769514516/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
dataclasses @ file:///tmp/build/80754af9/dataclasses_1614363715916/work
google-auth @ file:///opt/conda/conda-bld/google-auth_1646735974934/work
google-auth-oauthlib==0.4.1
grpcio @ file:///C:/ci/grpcio_1613114536274/work
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
idna-ssl @ file:///C:/ci/idna_ssl_1611752657819/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1631916885413/work
kiwisolver @ file:///C:/ci/kiwisolver_1612282446297/work
Markdown @ file:///C:/ci/markdown_1614364003005/work
matplotlib @ file:///C:/ci/matplotlib-suite_1613408055530/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict @ file:///C:/ci/multidict_1607362067746/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1603480701039/work
oauthlib @ file:///tmp/abs_08ngfezid4/croots/recipe/oauthlib_1659642459222/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
pandas @ file:///C:/ci/pandas_1608056614942/work
Pillow @ file:///C:/ci/pillow_1625670739152/work
protobuf==3.17.2
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT @ file:///C:/ci/pyjwt_1619651879817/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
PySocks @ file:///C:/ci/pysocks_1605305839978/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard @ file:///tmp/build/80754af9/tensorboard_1633093581375/work/tensorboard-2.6.0-py3-none-any.whl
tensorboard-data-server @ file:///C:/ci/tensorboard-data-server_1633035239634/work/tensorboard_data_server-0.6.0-py3-none-any.whl
tensorboard-plugin-wit==1.6.0
torch @ file:///C:/Users/11853/torch-1.8.2%2Bcu111-cp36-cp36m-win_amd64.whl
torchvision @ file:///C:/Users/11853/torchvision-0.9.2%2Bcu111-cp36-cp36m-win_amd64.whl
tornado @ file:///C:/ci/tornado_1606942379977/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306197271/work
wincertstore==0.2
yarl @ file:///C:/ci/yarl_1606921989557/work
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
